make select media as global function
